feat: implement CommonControls component for unified control buttons across platforms
