+++ view/desktop/src/parts/HeadBar/Controls/MacOSControls.tsx
@@ -2,6 +2,7 @@ import { useContext, type HTMLProps } from "react";
 
 import { cn } from "@/utils";
 
+import { CommonControls } from "./CommonControls";
 import { ControlButton } from "./ControlButton";
 import ControlsContext from "./ControlsContext";
 import { ControlsIcons } from "./icons";
@@ -13,25 +14,30 @@ export function MacOSControls({ className, ...props }: HTMLProps<HTMLDivElement>
   const { minimizeWindow, maximizeWindow, closeWindow } = useContext(ControlsContext);
 
   return (
-    <div className={cn("group flex gap-2 px-4 text-black", className)} {...props}>
-      <ControlButton
-        onClick={closeWindow}
-        className={cn(buttonStyles, "bg-[#ff544d] hover:bg-[#ff544d] active:bg-[#bf403a] dark:border-none")}
-      >
-        <ControlsIcons.closeMac className={cn(iconStyles)} />
-      </ControlButton>
-      <ControlButton
-        onClick={minimizeWindow}
-        className={cn(buttonStyles, "bg-[#ffbd2e] hover:bg-[#ffbd2e] active:bg-[#bf9122] dark:border-none")}
-      >
-        <ControlsIcons.minMac className={cn(iconStyles)} />
-      </ControlButton>
-      <ControlButton
-        onClick={maximizeWindow}
-        className={cn(buttonStyles, "bg-[#28c93f] hover:bg-[#28c93f] active:bg-[#1e9930] dark:border-none")}
-      >
-        <ControlsIcons.fullMac className={cn(iconStyles)} />
-      </ControlButton>
+    <div className={cn("group flex h-full justify-between", className)} {...props}>
+      <div className="flex gap-2 px-4 text-black">
+        <ControlButton
+          onClick={closeWindow}
+          className={cn(buttonStyles, "bg-[#ff544d] hover:bg-[#ff544d] active:bg-[#bf403a] dark:border-none")}
+        >
+          <ControlsIcons.closeMac className={cn(iconStyles)} />
+        </ControlButton>
+        <ControlButton
+          onClick={minimizeWindow}
+          className={cn(buttonStyles, "bg-[#ffbd2e] hover:bg-[#ffbd2e] active:bg-[#bf9122] dark:border-none")}
+        >
+          <ControlsIcons.minMac className={cn(iconStyles)} />
+        </ControlButton>
+        <ControlButton
+          onClick={maximizeWindow}
+          className={cn(buttonStyles, "bg-[#28c93f] hover:bg-[#28c93f] active:bg-[#1e9930] dark:border-none")}
+        >
+          <ControlsIcons.fullMac className={cn(iconStyles)} />
+        </ControlButton>
+      </div>
+      <div className="flex h-full">
+        <CommonControls />
+      </div>
     </div>
   );
 }
