+++ view/desktop/src/parts/HeadBar/Controls/WindowsControls.tsx
@@ -1,9 +1,8 @@
 import { useContext, type HTMLProps } from "react";
-import { useTranslation } from "react-i18next";
 
-import { useTabbedPaneStore } from "@/store/tabbedPane";
 import { cn } from "@/utils";
 
+import { CommonControls } from "./CommonControls";
 import { ControlButton } from "./ControlButton";
 import ControlsContext from "./ControlsContext";
 import { ControlsIcons } from "./icons";
@@ -13,53 +12,10 @@ import { ControlsIcons } from "./icons";
 
 export function WindowsControls({ className, ...props }: HTMLProps<HTMLDivElement>) {
   const { isWindowMaximized, minimizeWindow, maximizeWindow, closeWindow } = useContext(ControlsContext);
-  const { showDebugPanels, setShowDebugPanels } = useTabbedPaneStore();
-  const { t } = useTranslation(["ns1", "ns2"]);
-
-  const onAddPanel = (type?: string) => {
-    const api = useTabbedPaneStore.getState().api;
-    if (type && api?.getPanel(type) !== undefined) {
-      api.getPanel(type)?.focus();
-      return;
-    }
-
-    api?.addPanel({
-      id: type && type !== "nested" ? type : `id_${Date.now().toString()}`,
-      component: type ?? "Default",
-      title: type ?? `Tab ${Date.now()}`,
-      renderer: "onlyWhenVisible",
-    });
-  };
 
   return (
     <div className={cn("flex h-full", className)} {...props}>
-      <ControlButton
-        onClick={() => onAddPanel("Home")}
-        className="active:background-(--moss-button-icon-color)/[.03] h-full w-[46px] cursor-default rounded-none bg-transparent text-(--moss-button-icon-color)/90 hover:bg-[#0000000d]"
-        title={t("home")}
-      >
-        <span className="material-symbols-outlined">home</span>
-      </ControlButton>
-      <ControlButton
-        onClick={() => onAddPanel("Settings")}
-        className="active:background-(--moss-button-icon-color)/[.03] h-full w-[46px] cursor-default rounded-none bg-transparent text-(--moss-button-icon-color)/90 hover:bg-[#0000000d]"
-        title={t("settings")}
-      >
-        <span className="material-symbols-outlined">settings</span>
-      </ControlButton>
-      <ControlButton
-        onClick={() => onAddPanel("Logs")}
-        className="active:background-(--moss-button-icon-color)/[.03] h-full w-[46px] cursor-default rounded-none bg-transparent text-(--moss-button-icon-color)/90 hover:bg-[#0000000d]"
-        title={t("logs")}
-      >
-        <span className="material-symbols-outlined">terminal</span>
-      </ControlButton>
-      <ControlButton
-        onClick={() => setShowDebugPanels(!showDebugPanels)}
-        className="active:background-(--moss-button-icon-color)/[.03] h-full w-[46px] cursor-default rounded-none bg-transparent text-(--moss-button-icon-color)/90 hover:bg-[#0000000d]"
-      >
-        <span className="material-symbols-outlined">{showDebugPanels ? "hide_source" : "pest_control"}</span>
-      </ControlButton>
+      <CommonControls />
       <ControlButton
         onClick={minimizeWindow}
         className="active:background-(--moss-button-icon-color)/[.03] h-full w-[46px] cursor-default rounded-none bg-transparent text-(--moss-button-icon-color)/90 hover:bg-[#0000000d]"
